Creating Hud of Coins and improving Nim damage

+++ Project2/Assets/Scripts/Enemy Behaivers/Stinger.cs
@@ -65,7 +65,7 @@ public class Stinger : Enemy
             //Vector3 position = transform.position + Quaternion.Euler(0, 0, z) * Vector3.right / 3f;
 			bullet.transform.position = transform.position + direction.normalized * 0.36f;
             bullet.direction = direction;
-			bullet.damage = damage;
+			bullet.damage = (int)damage;
             bullet.speed = bulletSpeed;
         }
         
@@ -103,7 +103,7 @@ public class Stinger : Enemy
                 Vector3 position = transform.position + Quaternion.Euler(0, 0, z) * Vector3.right / 3f;
 				bullet.transform.position = transform.position + direction.normalized * 0.36f;
 				bullet.direction = direction;
-                bullet.damage = damage;
+                bullet.damage = (int) damage;
                 bullet.speed = bulletSpeed;
             }
         }
